third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/row-auto-repeat-auto-016-expected.txt

This is a testharness.js-based test.
[FAIL] Auto-fit repeat tracks with auto size and auto/explicit item placement
  assert_equals: expected "200px 200px 200px 200px 0px 200px 0px 0px 0px 0px" but got "none"
Harness: the test ran to completion.
